requirements.txt
streamlit
pytesseract
pdf2image
pdfplumber
psycopg2
Pillow
google-generativeai
opencv-python
numpy